# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /graph/entity_identifiers:
    get:
      summary: "get the graph entity identifiers"
      tags:
        - "graph"
        - "graph entity identifier"
      parameters:
        - name: search
          in: query
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entity identifiers"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntityIdentifier"
        "400":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResult"
  /graph/action/{action_id}:
    get:
      summary: "get the action graph entities"
      tags:
        - "graph"
        - "action"
      parameters:
        - name: action_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/dataset/{dataset_id}:
    get:
      summary: "get the dataset graph entities"
      tags:
        - "graph"
        - "dataset"
      parameters:
        - name: dataset_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/datastore/{datastore_id}:
    get:
      summary: "get the datastore graph entities"
      tags:
        - "graph"
        - "datastore"
      parameters:
        - name: datastore_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/event/{event_id}:
    get:
      summary: "get the event graph entities"
      tags:
        - "graph"
        - "event"
      parameters:
        - name: event_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/subscription/{subscription_id}:
    get:
      summary: "get the subscription graph entities"
      tags:
        - "graph"
        - "subscription"
      parameters:
        - name: subscription_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/trigger/{trigger_id}:
    get:
      summary: "get the trigger graph entities"
      tags:
        - "graph"
        - "trigger"
      parameters:
        - name: trigger_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/workflow/{workflow_id}:
    get:
      summary: "get the workflow graph entities"
      tags:
        - "graph"
        - "workflow"
      parameters:
        - name: workflow_id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/GraphEntity"
  /graph/sub_graph:
    get:
      summary: "get the sub graphs"
      tags:
        - "graph"
        - "subgraph"
      parameters:
        - name: related_type
          in: query
          type: string
        - name: related_id
          in: query
          type: string
        - name: engine_name
          in: query
          type: string
      responses:
        "200":
          description: "returns the sub graphs"
          schema:
            type: object
            properties:
              results:
                type: object
                #$ref: "#/definitions/SubGraph"
        "404":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResult"
    post:
      summary: "create a new sub graph"
      tags:
        - "graph"
      parameters:
        - name: debug
          in: query
          type: boolean
        - name: graph_entities
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/GraphEntity"
      responses:
        "200":
          description: "returns the graph entities"
          schema:
            type: object
            properties:
              results:
                type: object
                #$ref: "#/definitions/GraphEntity"
        "400":
          description: "error occurred"
          schema:
            $ref: "#/definitions/ErrorResult"
...
  /schema/engine/{engine_name}/datastore:
    get:
      summary: "get the datastore JSON Schema for the named engine"
      tags:
        - "schema"
        - "engine"
        - "datastore"
      parameters:
        - name: engine_name
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/JSONSchema"
        "400":
          description: "if engine is unknown"
          schema:
            $ref: "#/definitions/ErrorResult"
  /schema/datastore:
    get:
      summary: "get the datastore JSON Schema"
      tags:
        - "schema"
        - "datastore"
      parameters:
        - name: datastore_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/JSONSchema"
        "404":
          description: "if the optionally specified datastore could not be found"
          schema:
            $ref: "#/definitions/ErrorResult"
  /schema/action:
    get:
      summary: "get the action JSON Schema"
      tags:
        - "schema"
        - "action"
      parameters:
        - name: action_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/JSONSchema"
        "404":
          description: "if the optionally specified action could not be found"
          schema:
            $ref: "#/definitions/ErrorResult"
  /schema/action/{action_name}:
    get:
      summary: "get the action JSON Schema by action name"
      tags:
        - "schema"
        - "action"
      parameters:
        - name: action_name
          in: path
          type: string
          required: true
        - name: action_id
          in: query
          type: string
        - name: datastore_id
          in: query
          type: string
        - name: workflow_id
          in: query
          type: string
        - name: engine_name
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/JSONSchema"
        "400":
          description: "if the optionally specified action is unknown"
          schema:
            $ref: "#/definitions/ErrorResult"
        "404":
          description: "if the optionally specified action could not be found"
          schema:
            $ref: "#/definitions/ErrorResult"
...
  /schema/trigger:
    get:
      summary: "get the trigger JSON schema"
      tags:
        - "schema"
        - "trigger"
      parameters:
        - name: trigger_type_name
          in: query
          type: string
        - name: trigger_id
          in: query
          type: string
      responses:
        "200":
          description: "returns the JSON Schema"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/JSONSchema"
        "400":
          description: "if the optionally specified trigger is unknown"
          schema:
            $ref: "#/definitions/ErrorResult"
...
definitions:
...
  ErrorResult:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  GraphEntity:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
      name:
        type: string
      related_id:
        type: string
      related_is_a:
        type: string
      related_type:
        type: string
      state:
        type: string
      sub_type:
        type: string
    type: object
  GraphEntityIdentifier:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
      name:
        type: string
    type: object
...
  JSONSchema:
    description: JSON Schema object as defined by http://json-schema.org/schema
    type: object
...
  SubGraph:
    type: object
...